Reverted login to email
Now on login it asks for email, if i have time i will return and make it username but I simply dont have time to keep working on it. I created the html for the page to create the trading cards but running into issues. Need to merge this branch though as the issue for this branch is complete