Implemented a bunch of tests
Added a ton of tests to test models, views, and controllers to ensure coverage. Coverage is at about 50% which is not what I wanted but a lot of missing coverage is for files that were deafult in the framework that I did not touch. If i had more time i would fix this but I am running out of time